+++ AdventOfCode/Utils.cs
@@ -8,12 +8,12 @@ namespace AdventOfCode
     {
         public const string INPUT_DIR = @"input";
 
-        public static int[] GetInput(string filename, string delimiter)
+        public static long[] GetInput(string filename, string delimiter)
         {
             string path     = Path.Combine(INPUT_DIR, filename);
             string text     = File.ReadAllText(path);
             var input       = text.Split(delimiter).Where(x => double.TryParse(x, out _));
-            var parsedInput = input.Select(x => int.Parse(x));
+            var parsedInput = input.Select(x => long.Parse(x));
 
             return parsedInput.ToArray();
         }
